+++ sdk/core/azure-core/inc/azure/core/http/curl_transport.hpp
@@ -14,13 +14,6 @@
 #include "azure/core/http/transport.hpp"
 #include "azure/core/platform.hpp"
 
-#if defined(AZ_PLATFORM_LINUX)
-#include <openssl/opensslv.h>
-#if OPENSSL_VERSION_NUMBER >= 0x00905100L
-#define _azure_SUPPORT_SETTING_CAPATH
-#endif // OPENSSL_VERSION_NUMBER >= 0x00905100L
-#endif // defined(AZ_PLATFORM_LINUX)
-
 namespace Azure { namespace Core { namespace Http {
   class CurlNetworkConnection;
 
@@ -130,7 +123,7 @@ namespace Azure { namespace Core { namespace Http {
      */
     std::string CAInfo;
 
-#if defined(_azure_SUPPORT_SETTING_CAPATH)
+#if defined(AZ_PLATFORM_LINUX)
     /**
      * @brief Path to a directory which holds PEM encoded file, containing the certificate
      * authorities sent to libcurl handle directly.