+++ sdk/storage/azure-storage-files-datalake/src/datalake_options.cpp
@@ -3,8 +3,8 @@
 
 #include "azure/storage/files/datalake/datalake_options.hpp"
 
-namespace Azure { namespace Storage { namespace Files { namespace DataLake { namespace Models {
+namespace Azure { namespace Storage { namespace Files { namespace DataLake {
 
-  const DataLakeAudience DataLakeAudience::PublicAudience(Azure::Storage::_internal::StorageScope);
+  const DataLakeAudience DataLakeAudience::DefaultAudience(_internal::StorageDefaultAudience);
 
-}}}}} // namespace Azure::Storage::Files::DataLake::Models
+}}}} // namespace Azure::Storage::Files::DataLake