+++ sdk/storage/azure-storage-files-datalake/test/ut/datalake_path_client_test.cpp
@@ -489,22 +489,36 @@ namespace Azure { namespace Storage { namespace Test {
         InitStorageClientOptions<Azure::Identity::ClientSecretCredentialOptions>());
     auto clientOptions = InitStorageClientOptions<Files::DataLake::DataLakeClientOptions>();
 
-    // default audience
+    // audience by default
     auto pathClient
+        = Files::DataLake::DataLakePathClient(m_pathClient->GetUrl(), credential, clientOptions);
+    EXPECT_NO_THROW(pathClient.GetProperties());
+
+    // default audience
+    clientOptions.Audience = Files::DataLake::DataLakeAudience::DefaultAudience;
+    pathClient
+        = Files::DataLake::DataLakePathClient(m_pathClient->GetUrl(), credential, clientOptions);
+    EXPECT_NO_THROW(pathClient.GetProperties());
+
+    // service audience
+    auto keyCredential = _internal::ParseConnectionString(AdlsGen2ConnectionString()).KeyCredential;
+    auto accountName = keyCredential->AccountName;
+    clientOptions.Audience
+        = Files::DataLake::DataLakeAudience::CreateDataLakeServiceAccountAudience(accountName);
+    pathClient
         = Files::DataLake::DataLakePathClient(m_pathClient->GetUrl(), credential, clientOptions);
     EXPECT_NO_THROW(pathClient.GetProperties());
 
     // custom audience
     auto pathUrl = Azure::Core::Url(pathClient.GetUrl());
-    clientOptions.Audience = Files::DataLake::Models::DataLakeAudience(
-        pathUrl.GetScheme() + "://" + pathUrl.GetHost() + "/.default");
+    clientOptions.Audience
+        = Files::DataLake::DataLakeAudience(pathUrl.GetScheme() + "://" + pathUrl.GetHost());
     pathClient
         = Files::DataLake::DataLakePathClient(m_pathClient->GetUrl(), credential, clientOptions);
     EXPECT_NO_THROW(pathClient.GetProperties());
 
     // error audience
-    clientOptions.Audience
-        = Files::DataLake::Models::DataLakeAudience("https://disk.compute.azure.com/.default");
+    clientOptions.Audience = Files::DataLake::DataLakeAudience("https://disk.compute.azure.com");
     pathClient
         = Files::DataLake::DataLakePathClient(m_pathClient->GetUrl(), credential, clientOptions);
     EXPECT_THROW(pathClient.GetProperties(), StorageException);