+++ sdk/eventhubs/azure-messaging-eventhubs-checkpointstore-blob/CHANGELOG.md
@@ -1,13 +1,18 @@
 # Release History
 
-## 1.0.0-beta.3 (Unreleased)
+## 1.0.0-beta.2 (Unreleased)
 
 ### Features Added
-
-- Initial release. Split from the `azure-messaging-eventubs` package.
 
 ### Breaking Changes
 
 ### Bugs Fixed
 
 ### Other Changes
+
+## 1.0.0-beta.1 (2023-10-10)
+
+### Features Added
+
+- Initial release. Split from the `azure-messaging-eventubs-cpp` package.
+